#115 - Move class to another package
