+++ src/main/java/com/artipie/helm/misc/EmptyIndex.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.artipie.helm.metadata;
+package com.artipie.helm.misc;
 
 import com.artipie.asto.Content;
 import com.artipie.helm.misc.DateTimeNow;
